requirements.txt
-i https://pypi.org/simple
arrow==1.2.3 ; python_version >= '3.6'
attrs==23.1.0 ; python_version >= '3.7'
beautifulsoup4==4.12.2 ; python_full_version >= '3.6.0'
bs4==0.0.1
cachetools==5.3.0 ; python_version ~= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
google-api-core==2.11.0 ; python_version >= '3.7'
google-api-python-client==2.86.0 ; python_version >= '3.7'
google-auth==2.17.3 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0 ; python_version >= '3.7'
httplib2==0.22.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ics==0.7.2
idna==3.4 ; python_version >= '3.5'
parsedatetime==2.6
protobuf==4.23.0 ; python_version >= '3.7'
pyasn1==0.5.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyparsing==3.0.9 ; python_version >= '3.1'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyyaml==6.0
-e git+https://github.com/gmfrasca/rec-league-parser@5a9b5d2f209bb1829b2fa42449f7ffdb7297f6d4#egg=recleagueparser
requests==2.30.0 ; python_version >= '3.7'
rsa==4.9 ; python_version >= '3.6'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.4.1 ; python_version >= '3.7'
tatsu==5.8.3 ; python_version >= '3.8'
tzdata==2023.3 ; python_version >= '3.6'
tzlocal==4.3
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==2.0.2 ; python_version >= '3.7'
